Modify file structure for performance rating and manifest file
